% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-confidence-intervals.R
\name{ci_hi}
\alias{ci_hi}
\title{Confidence Interval Generic}
\usage{
ci_hi(.x, .na_rm = FALSE)
}
\arguments{
\item{.x}{A vector of numeric values}

\item{.na_rm}{A Boolean, defaults to FALSE. Passed to the quantile function.}
}
\value{
A numeric value.
}
\description{
Gets the upper 97.5\% quantile of a numeric vector.
}
\details{
Gets the upper 97.5\% quantile of a numeric vector.
}
\examples{
x <- mtcars$mpg
ci_hi(x)

}
\seealso{
Other Statistic: 
\code{\link{ci_lo}()},
\code{\link{tidy_kurtosis_vec}()},
\code{\link{tidy_range_statistic}()},
\code{\link{tidy_skewness_vec}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Statistic}
