% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-color-blind.R
\name{color_blind}
\alias{color_blind}
\title{Provide Colorblind Compliant Colors}
\usage{
color_blind()
}
\description{
Provide Colorblind Compliant Colors
}
\concept{Color Blind

8 Hex RGB color definitions suitable for charts for colorblind people.}
