% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-tidy-gamma.R
\name{tidy_gamma}
\alias{tidy_gamma}
\title{Tidy Randomly Generated Gaussian Tibble}
\usage{
tidy_gamma(.n = 50, .shape = 1, .rate = 1, .num_sims = 1)
}
\arguments{
\item{.n}{The number of randomly generated points you want.}

\item{.shape}{The mean of the randomly generated data. This is strictly 0 to infinity.}

\item{.rate}{The standard deviation of the randomly generated data. This is
strictly from 0 to infinity.}

\item{.num_sims}{The number of randomly generated simulations you want.}
}
\value{
A tibble of randomly generated data.
}
\description{
This function will generate \code{n} random points from a gamma
distribution with a user provided, shape, rate, and number of
random simulations to be produced. The function returns a tibble with the
simulation number column the x column which corresponds to the n randomly
generated points, the \code{d_}, \code{p_} and \code{q_} data points as well.

The data is returned un-grouped.
}
\details{
This function uses the underlying \code{stats::rgamma()}, and its underlying
\code{p}, \code{d}, and \code{q} functions.
}
\examples{
tidy_gamma()

}
\seealso{
\url{hhttps://www.statology.org/fit-gamma-distribution-to-dataset-in-r/}

\url{https://en.wikipedia.org/wiki/Gamma_distribution}

Other Data Generator: 
\code{\link{tidy_normal}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Data Generator}
