% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-tidy-rnorm.R
\name{tidy_normal}
\alias{tidy_normal}
\title{Tidy Randomly Generated Gaussian Tibble}
\usage{
tidy_normal(.n = 50, .mean = 0, .sd = 1, .num_sims = 1)
}
\arguments{
\item{.n}{The number of randomly generated points you want.}

\item{.mean}{The mean of the randomly generated data.}

\item{.sd}{The standard deviation of the randomly generated data.}

\item{.num_sims}{The number of randomly generated simulations you want.}
}
\value{
A tibble of randomly generated.
}
\description{
This function will generate \code{n} randong points from a gaussian
distribution with a user provided, mean, standard deviation and number of
random simulations to be produced. The function returns a tibble with the
simulation number column the x column which corresponds to the n randomly
generated points, the \code{dnorm}, \code{pnorm} and \code{qnorm} data points as well.

The data is returned un-grouped.
}
\details{
This function uses the underlying \code{stats::rnorm()}, \code{stats::pnorm()},
and \code{stats::qnorm()} functions to generate data from the given parameters.
}
\examples{
tidy_normal()

}
\seealso{
\url{https://data-flair.training/blogs/normal-distribution-in-r/}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Data Generator}
