% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-tidy-poisson.R
\name{tidy_poisson}
\alias{tidy_poisson}
\title{Tidy Randomly Generated Poisson Distribution Tibble}
\usage{
tidy_poisson(.n = 50, .lambda = 1, .num_sims = 1)
}
\arguments{
\item{.n}{The number of randomly generated points you want.}

\item{.lambda}{A vector of non-negative means.}

\item{.num_sims}{The number of randomly generated simulations you want.}
}
\value{
A tibble of randomly generated data.
}
\description{
This function will generate \code{n} random points from a Poisson
distribution with a user provided, \code{.lambda}, and number of
random simulations to be produced. The function returns a tibble with the
simulation number column the x column which corresponds to the n randomly
generated points, the \code{d_}, \code{p_} and \code{q_} data points as well.

The data is returned un-grouped.

The columns that are output are:
\itemize{
\item \code{sim_number} The current simulation number.
\item \code{x} The current value of \code{n} for the current simulation.
\item \code{y} The randomly generated data point.
\item \code{dx} The \code{x} value from the \code{\link[stats:density]{stats::density()}} function.
\item \code{dy} The \code{y} value from the \code{\link[stats:density]{stats::density()}} function.
\item \code{p} The values from the resulting p_ function of the distribution family.
\item \code{q} The values from the resulting q_ function of the distribution family.
}
}
\details{
This function uses the underlying \code{stats::rpois()}, and its underlying
\code{p}, \code{d}, and \code{q} functions. For more information please see \code{\link[stats:Poisson]{stats::rpois()}}
}
\examples{
tidy_poisson()
}
\seealso{
\url{https://r-coder.com/poisson-distribution-r/}

\url{https://en.wikipedia.org/wiki/Poisson_distribution}

Other Poisson: 
\code{\link{tidy_zero_truncated_poisson}()},
\code{\link{util_poisson_param_estimate}()}

Other Discrete Distribution: 
\code{\link{tidy_binomial}()},
\code{\link{tidy_hypergeometric}()},
\code{\link{tidy_negative_binomial}()},
\code{\link{tidy_zero_truncated_binomial}()},
\code{\link{tidy_zero_truncated_negative_binomial}()},
\code{\link{tidy_zero_truncated_poisson}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Discrete Distribution}
\concept{Poisson}
