% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-range.R
\name{tidy_range_statistic}
\alias{tidy_range_statistic}
\title{Get the range statistic}
\usage{
tidy_range_statistic(.x)
}
\arguments{
\item{.x}{A numeric vector}
}
\value{
A single number, the range statistic
}
\description{
Takes in a numeric vector and returns back the range of that
vector
}
\details{
Takes in a numeric vector and returns the range of that vector using
the \code{diff} and \code{range} functions.
}
\examples{
tidy_range_statistic(seq(1:10))

}
\seealso{
Other Statistic: 
\code{\link{tidy_kurtosis_vec}()},
\code{\link{tidy_skewness_vec}()}
}
\author{
Steven P. Sandeson II, MPH
}
\concept{Statistic}
