% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est-param-binomial.R
\name{util_binomial_param_estimate}
\alias{util_binomial_param_estimate}
\title{Estimate Binomial Parameters}
\usage{
util_binomial_param_estimate(.x, .size = NULL, .auto_gen_empirical = TRUE)
}
\arguments{
\item{.x}{The vector of data to be passed to the function. Must be numeric, and
all values must be 0 <= x <= 1}

\item{.size}{Number of trials, zero or more.}

\item{.auto_gen_empirical}{This is a boolean value of TRUE/FALSE with default
set to TRUE. This will automatically create the \code{tidy_empirical()} output
for the \code{.x} parameter and use the \code{tidy_combine_distributions()}. The user
can then plot out the data using \verb{$combined_data_tbl} from the function output.}
}
\value{
A tibble/list
}
\description{
This function will check to see if some given vector \code{.x} is
either a numeric vector or a factor vector with at least two levels then it
will cause an error and the function will abort.
}
\details{
This function will attempt to estimate the binomial p_hat and size
parameters given some vector of values.
}
\examples{
library(dplyr)
library(ggplot2)

tb <- rbinom(50, 1, .1)
util_binomial_param_estimate(tb)$parameter_tbl
util_binomial_param_estimate(tb)$combined_data_tbl \%>\%
  ggplot(aes(x = y, group = dist_type, color = "black", fill = dist_type)) +
  geom_histogram(binwidth = 0.5) +
  theme_minimal() +
  theme(legend.position = "bottom")

}
\seealso{
Other Parameter Estimation: 
\code{\link{util_beta_param_estimate}()}

Other Binomial: 
\code{\link{tidy_binomial}()},
\code{\link{tidy_negative_binomial}()},
\code{\link{tidy_zero_truncated_binomial}()},
\code{\link{tidy_zero_truncated_negative_binomial}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Binomial}
\concept{Parameter Estimation}
