% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aic-chisq.R
\name{util_chisq_aic}
\alias{util_chisq_aic}
\title{Calculate Akaike Information Criterion (AIC) for Chi-Square Distribution}
\usage{
util_chisq_aic(.x)
}
\arguments{
\item{.x}{A numeric vector containing the data to be fitted to a chi-square distribution.}
}
\value{
The AIC value calculated based on the fitted chi-square distribution to the provided data.
}
\description{
This function estimates the parameters of a chi-square distribution from the provided data using maximum likelihood estimation,
and then calculates the AIC value based on the fitted distribution.
}
\details{
This function calculates the Akaike Information Criterion (AIC) for a chi-square distribution fitted to the provided data.
}
\examples{
# Example 1: Calculate AIC for a sample dataset
data <- c(1, 2, 3, 4, 5)
util_chisq_aic(data)

}
\seealso{
Other Utility: 
\code{\link{check_duplicate_rows}()},
\code{\link{convert_to_ts}()},
\code{\link{quantile_normalize}()},
\code{\link{tidy_mcmc_sampling}()},
\code{\link{util_beta_aic}()},
\code{\link{util_normal_aic}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Utility}
