% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aic-f.R
\name{util_f_aic}
\alias{util_f_aic}
\title{Calculate Akaike Information Criterion (AIC) for F Distribution}
\usage{
util_f_aic(.x)
}
\arguments{
\item{.x}{A numeric vector containing the data to be fitted to an F
distribution.}
}
\value{
The AIC value calculated based on the fitted F distribution to the
provided data.
}
\description{
This function calculates the Akaike Information Criterion (AIC) for an F
distribution fitted to the provided data.
}
\details{
This function fits an F distribution to the input data using maximum
likelihood estimation and then computes the Akaike Information Criterion (AIC)
based on the fitted distribution.
}
\examples{
# Generate F-distributed data
set.seed(123)
x <- rf(100, df1 = 5, df2 = 10, ncp = 1)

# Calculate AIC for the generated data
util_f_aic(x)

}
\seealso{
\code{\link[stats]{rf}} for generating F-distributed data,
\code{\link[stats]{optim}} for optimization.
}
