% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-geometric-tbl.R
\name{util_geometric_stats_tbl}
\alias{util_geometric_stats_tbl}
\title{Distribution Statistics}
\usage{
util_geometric_stats_tbl(.data)
}
\arguments{
\item{.data}{The data being passed from a \code{tidy_} distribution function.}
}
\value{
A tibble
}
\description{
Returns distribution statistics in a tibble.
}
\details{
This function will take in a tibble and returns the statistics
of the given type of \code{tidy_} distribution. It is required that data be
passed from a \code{tidy_} distribution function.
}
\examples{
tidy_geometric() \%>\%
  util_geometric_stats_tbl()

}
\seealso{
Other Geometric: 
\code{\link{tidy_geometric}()},
\code{\link{tidy_zero_truncated_geometric}()},
\code{\link{util_geometric_param_estimate}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Geometric}
