% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est-param-paralogistic.R
\name{util_paralogistic_param_estimate}
\alias{util_paralogistic_param_estimate}
\title{Estimate Paralogistic Parameters}
\usage{
util_paralogistic_param_estimate(.x, .auto_gen_empirical = TRUE)
}
\arguments{
\item{.x}{The vector of data to be passed to the function.}

\item{.auto_gen_empirical}{This is a boolean value of TRUE/FALSE with default
set to TRUE. This will automatically create the \code{tidy_empirical()} output
for the \code{.x} parameter and use the \code{tidy_combine_distributions()}. The user
can then plot out the data using \verb{$combined_data_tbl} from the function output.}
}
\value{
A tibble/list
}
\description{
The function will return a list output by default, and if the parameter
\code{.auto_gen_empirical} is set to \code{TRUE} then the empirical data given to the
parameter \code{.x} will be run through the \code{tidy_empirical()} function and combined
with the estimated paralogistic data.

The method of parameter estimation is:
\itemize{
\item MLE
}
}
\details{
This function will attempt to estimate the paralogistic shape and rate
parameters given some vector of values.
}
\examples{
library(dplyr)
library(ggplot2)

x <- mtcars$mpg
output <- util_paralogistic_param_estimate(x)

output$parameter_tbl

output$combined_data_tbl |>
  tidy_combined_autoplot()

t <- tidy_paralogistic(50, 2.5, 1.4)[["y"]]
util_paralogistic_param_estimate(t)$parameter_tbl

}
\seealso{
Other Parameter Estimation: 
\code{\link{util_bernoulli_param_estimate}()},
\code{\link{util_beta_param_estimate}()},
\code{\link{util_binomial_param_estimate}()},
\code{\link{util_burr_param_estimate}()},
\code{\link{util_cauchy_param_estimate}()},
\code{\link{util_chisquare_param_estimate}()},
\code{\link{util_exponential_param_estimate}()},
\code{\link{util_f_param_estimate}()},
\code{\link{util_gamma_param_estimate}()},
\code{\link{util_generalized_beta_param_estimate}()},
\code{\link{util_generalized_pareto_param_estimate}()},
\code{\link{util_geometric_param_estimate}()},
\code{\link{util_hypergeometric_param_estimate}()},
\code{\link{util_inverse_burr_param_estimate}()},
\code{\link{util_inverse_pareto_param_estimate}()},
\code{\link{util_inverse_weibull_param_estimate}()},
\code{\link{util_logistic_param_estimate}()},
\code{\link{util_lognormal_param_estimate}()},
\code{\link{util_negative_binomial_param_estimate}()},
\code{\link{util_normal_param_estimate}()},
\code{\link{util_pareto1_param_estimate}()},
\code{\link{util_pareto_param_estimate}()},
\code{\link{util_poisson_param_estimate}()},
\code{\link{util_t_param_estimate}()},
\code{\link{util_triangular_param_estimate}()},
\code{\link{util_uniform_param_estimate}()},
\code{\link{util_weibull_param_estimate}()},
\code{\link{util_zero_truncated_binomial_param_estimate}()},
\code{\link{util_zero_truncated_geometric_param_estimate}()},
\code{\link{util_zero_truncated_negative_binomial_param_estimate}()},
\code{\link{util_zero_truncated_poisson_param_estimate}()}

Other Paralogistic: 
\code{\link{util_paralogistic_aic}()},
\code{\link{util_paralogistic_stats_tbl}()}
}
\concept{Paralogistic}
\concept{Parameter Estimation}
