% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aic-weibull.R
\name{util_weibull_aic}
\alias{util_weibull_aic}
\title{Calculate Akaike Information Criterion (AIC) for Weibull Distribution}
\usage{
util_weibull_aic(.x)
}
\arguments{
\item{.x}{A numeric vector containing the data to be fitted to a Weibull distribution.}
}
\value{
The AIC value calculated based on the fitted Weibull distribution to the provided data.
}
\description{
This function estimates the shape and scale parameters of a Weibull distribution
from the provided data using maximum likelihood estimation,
and then calculates the AIC value based on the fitted distribution.
}
\details{
This function calculates the Akaike Information Criterion (AIC) for a Weibull
distribution fitted to the provided data.

This function fits a Weibull distribution to the provided data using maximum
likelihood estimation. It estimates the shape and scale parameters
of the Weibull distribution using maximum likelihood estimation. Then, it
calculates the AIC value based on the fitted distribution.

Initial parameter estimates: The function uses the method of moments estimates
as starting points for the shape and scale parameters of the Weibull
distribution.

Optimization method: The function uses the optim function for optimization.
You might explore different optimization methods within optim for potentially
better performance.

Goodness-of-fit: While AIC is a useful metric for model comparison, it's recommended
to also assess the goodness-of-fit of the chosen model using visualization
and other statistical tests.
}
\examples{
# Example 1: Calculate AIC for a sample dataset
set.seed(123)
x <- rweibull(100, shape = 2, scale = 1)
util_weibull_aic(x)

}
\seealso{
Other Utility: 
\code{\link{check_duplicate_rows}()},
\code{\link{convert_to_ts}()},
\code{\link{quantile_normalize}()},
\code{\link{tidy_mcmc_sampling}()},
\code{\link{util_beta_aic}()},
\code{\link{util_binomial_aic}()},
\code{\link{util_cauchy_aic}()},
\code{\link{util_chisq_aic}()},
\code{\link{util_exponential_aic}()},
\code{\link{util_gamma_aic}()},
\code{\link{util_geometric_aic}()},
\code{\link{util_hypergeometric_aic}()},
\code{\link{util_logistic_aic}()},
\code{\link{util_lognormal_aic}()},
\code{\link{util_normal_aic}()},
\code{\link{util_pareto_aic}()},
\code{\link{util_poisson_aic}()},
\code{\link{util_uniform_aic}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Utility}
