% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-color-blind.R
\name{td_scale_color_colorblind}
\alias{td_scale_color_colorblind}
\title{Provide Colorblind Compliant Colors}
\usage{
td_scale_color_colorblind(..., theme = "td")
}
\arguments{
\item{theme}{This defaults to \code{td} and that is the only allowed value}
}
\description{
Provide Colorblind Compliant Colors
}
\seealso{
Other ColorBlind: 
\code{\link{td_scale_fill_colorblind}()}
}
\concept{ColorBlind}
