% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-tidy-gamma.R
\name{tidy_gamma}
\alias{tidy_gamma}
\title{Tidy Randomly Generated Gamma Tibble}
\usage{
tidy_gamma(.n = 50, .shape = 1, .rate = 1, .num_sims = 1)
}
\arguments{
\item{.n}{The number of randomly generated points you want.}

\item{.shape}{This is strictly 0 to infinity.}

\item{.rate}{The standard deviation of the randomly generated data. This is
strictly from 0 to infinity.}

\item{.num_sims}{The number of randomly generated simulations you want.}
}
\value{
A tibble of randomly generated data.
}
\description{
This function will generate \code{n} random points from a gamma
distribution with a user provided, \code{.shape}, \code{.rate}, and number of
random simulations to be produced. The function returns a tibble with the
simulation number column the x column which corresponds to the n randomly
generated points, the \code{d_}, \code{p_} and \code{q_} data points as well.

The data is returned un-grouped.

The columns that are output are:
\itemize{
\item \code{sim_number} The current simulation number.
\item \code{x} The current value of \code{n} for the current simulation.
\item \code{y} The randomly generated data point.
\item \code{dx} The \code{x} value from the \code{\link[stats:density]{stats::density()}} function.
\item \code{dy} The \code{y} value from the \code{\link[stats:density]{stats::density()}} function.
\item \code{p} The values from the resulting p_ function of the distribution family.
\item \code{q} The values from the resulting q_ function of the distribution family.
}
}
\details{
This function uses the underlying \code{stats::rgamma()}, and its underlying
\code{p}, \code{d}, and \code{q} functions. For more information please see \code{\link[stats:GammaDist]{stats::rgamma()}}
}
\examples{
tidy_gamma()

}
\seealso{
\url{https://www.statology.org/fit-gamma-distribution-to-dataset-in-r/}

\url{https://en.wikipedia.org/wiki/Gamma_distribution}

Other Data Generator: 
\code{\link{tidy_beta}()},
\code{\link{tidy_cauchy}()},
\code{\link{tidy_chisquare}()},
\code{\link{tidy_exponential}()},
\code{\link{tidy_f}()},
\code{\link{tidy_hypergeometric}()},
\code{\link{tidy_logistic}()},
\code{\link{tidy_lognormal}()},
\code{\link{tidy_normal}()},
\code{\link{tidy_poisson}()},
\code{\link{tidy_uniform}()},
\code{\link{tidy_weibull}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Data Generator}
