% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-tidy-rnorm.R
\name{tidy_rnorm}
\alias{tidy_rnorm}
\title{Tidy Randomly Generated Gaussian Tibble}
\usage{
tidy_rnorm(.n = 50, .mean = 0, .sd = 1, .num_sims = 1)
}
\arguments{
\item{.n}{The number of randomly generated points you want.}

\item{.mean}{The mean of the randomly generated data.}

\item{.sd}{The standard deviation of the randomly generated data.}

\item{.num_sims}{The number of randomly generated simulations you want.}
}
\value{
A tibble of randomly generated.
}
\description{

}
\details{
This function uses the underlying \code{stats::rnorm()} function to generate
data from the given parameters.
}
\examples{
tidy_rnorm()

}
\author{
Steven P. Sanderson II, MPH
}
\concept{Data Generator}
