% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est-param-beta.R
\name{util_beta_param_estimate}
\alias{util_beta_param_estimate}
\title{Estimate Beta Parameters}
\usage{
util_beta_param_estimate(.x, .auto_gen_empirical = TRUE)
}
\arguments{
\item{.x}{The vector of data to be passed to the function. Must be numeric, and
all values must be 0 <= x <= 1}

\item{.auto_gen_empirical}{This is a boolean value of TRUE/FALSE with default
set to TRUE. This will automatically create the \code{tidy_empirical()} output
for the \code{.x} parameter and use the \code{tidy_combine_distributions()}. The user
can then plot out the data using \verb{$combined_data_tbl} from the function output.}
}
\value{
A tibble
}
\description{
This function is a wrapper for the \code{\link[EnvStats:ebeta]{EnvStats::ebeta()}} function.
It will automatically scale the data from 0 to 1 if it is not already. This means
you can pass a vector like \code{mtcars$mpg} and not worry about it.
}
\details{
This is a wrapper for the \code{\link[EnvStats:ebeta]{EnvStats::ebeta()}} function.
}
\examples{
library(dplyr)

tb <- tidy_beta(.n = 50, .shape1 = 2.5, .shape2 = 1.4, .ncp = 0) \%>\%
  pull(y)
util_beta_param_estimate(tb)

}
\seealso{
Other Beta: 
\code{\link{tidy_beta}()},
\code{\link{tidy_generalized_beta}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Beta}
\concept{Parameter Estimation}
